Index a file for faster access to parts of the file

Index a file for faster access to parts of the file

make_file_index(
  pathname,
  skip = 0L,
  n_max = Inf,
  newline = "\n",
  bfr_size = 1e+07
)

save_file_index(index, file)

read_file_index(file)

Arguments

pathname
(character) The file to be indexed.

skip
(numeric) The number of bytes to skip before start indexing.

n_max
(numeric) The maximum number of bytes to scan.

newline
(character) The character to scan for.

bfr_size
(numeric) The number of bytes to read in each iteration.

index
(numeric vector) A sorted index of file byte positions.

file
A pathname to a *.index file to be created or read from.

Value

A numeric vector of file byte offsets that corresponds to the beginning of a line, i.e. a position in the file that was preceeded by a newline character. The first line is at file byte offset 0, which is also always the first element in the returned vector.

Examples

pathname <- system.file("exdata", "accounting", package = "wyntonquery")
index <- read_file_index(sprintf("%s.index", pathname))
cat(sprintf("Number of jobs: %d\n", length(index)))

#> Number of jobs: 1005

con <- open_file_at(pathname, position = index[1001])

## Get all jobs after the 1000:th job
jobs <- read_sge_accounting(con)
close(con)
print(jobs)

#> # A tibble: 4 x 45
#>   qname   hostname  group  owner  job_name           job_number account priority
#>   <chr>   <chr>     <chr>  <chr>  <chr>                   <int> <chr>      <int>
#> 1 short.q cin-hmid1 group5 owner… qss_process_resam…        425 sge           10
#> 2 short.q cin-hmid1 group5 owner… qss_process_resam…        425 sge           10
#> 3 short.q cin-hmid1 group5 owner… qss_process_resam…        425 sge           10
#> 4 short.q cin-hmid1 group5 owner… qss_process_resam…        425 sge           10
#> # … with 37 more variables: submission_time <dttm>, start_time <dttm>,
#> #   end_time <dttm>, failed <int>, exit_status <int>, ru_wallclock <drtn>,
#> #   ru_utime <drtn>, ru_stime <drtn>, ru_maxrss <chr>, ru_ixrss <chr>,
#> #   ru_ismrss <chr>, ru_idrss <chr>, ru_isrss <chr>, ru_minflt <dbl>,
#> #   ru_majflt <dbl>, ru_nswap <dbl>, ru_inblock <dbl>, ru_oublock <dbl>,
#> #   ru_msgsnd <dbl>, ru_msgrcv <dbl>, ru_nsignals <dbl>, ru_nvcsw <dbl>,
#> #   ru_nivcsw <dbl>, project <chr>, department <chr>, granted_pe <chr>,
#> #   slots <int>, task_number <int>, cpu <drtn>, mem <chr>, io <chr>,
#> #   category <chr>, iow <drtn>, pe_taskid <chr>, maxvmem <dbl>, arid <dbl>,
#> #   ar_sub_time <dttm>