Read a tab-delimited SGE accounting file (without parsing it)

Read a tab-delimited SGE accounting file (without parsing it)

read_raw_sge_accounting(file, skip = 4L, ...)

write_raw_sge_accounting(x, file, header = attr(x, "header"), ...)

anonymize_raw_sge_accounting(x, fields = c("group", "owner", "project"))

Arguments

file
(character) The SGE accounting file to read.

skip
(integer) Number of lines to skip before parsing file content.

x
(raw_sge_accounting) An tibble data frame of class raw_sge_accounting.

header
(character vector) Zero of more header lines to be written at the top of the file.

fields
(character vector) Columns to be anonymized.

...
(optional) Addition arguments passed to readr::read_delim().

Details

The SGE accounting file is typically located in a subfolder of the folder $SGE_ROOT/$SGE_CELL/. On Wynton HPC, the pathname is given by sge_accounting_file().

Examples

pathname <- system.file("exdata", "accounting", package = "wyntonquery")

jobs <- read_raw_sge_accounting(pathname)
print(jobs)

#> # A tibble: 1,000 x 45
#>    qname    hostname group  owner   job_name job_number account priority
#>    <chr>    <chr>    <chr>  <chr>   <chr>         <int> <chr>      <int>
#>  1 member.q cin-id3  group4 owner09 sleep.sh          1 sge            0
#>  2 long.q   cc-id2   group4 owner09 sleep.sh          2 sge           19
#>  3 long.q   cc-id2   group4 owner09 sleep.sh          2 sge           19
#>  4 long.q   cc-id2   group4 owner09 sleep.sh          2 sge           19
#>  5 long.q   cc-id2   group4 owner09 sleep.sh          2 sge           19
#>  6 long.q   cc-id2   group4 owner09 sleep.sh          2 sge           19
#>  7 long.q   cc-id2   group4 owner09 sleep.sh          2 sge           19
#>  8 member.q cin-id2  group4 owner09 sleep.sh          2 sge            0
#>  9 long.q   cin-id2  group4 owner09 sleep.sh          2 sge           19
#> 10 long.q   cin-id2  group4 owner09 sleep.sh          2 sge           19
#> # … with 990 more rows, and 37 more variables: submission_time <dbl>,
#> #   start_time <dbl>, end_time <dbl>, failed <int>, exit_status <int>,
#> #   ru_wallclock <dbl>, ru_utime <dbl>, ru_stime <dbl>, ru_maxrss <dbl>,
#> #   ru_ixrss <dbl>, ru_ismrss <dbl>, ru_idrss <dbl>, ru_isrss <dbl>,
#> #   ru_minflt <dbl>, ru_majflt <dbl>, ru_nswap <dbl>, ru_inblock <dbl>,
#> #   ru_oublock <dbl>, ru_msgsnd <dbl>, ru_msgrcv <dbl>, ru_nsignals <dbl>,
#> #   ru_nvcsw <dbl>, ru_nivcsw <dbl>, project <chr>, department <chr>,
#> #   granted_pe <chr>, slots <int>, task_number <int>, cpu <dbl>, mem <dbl>,
#> #   io <dbl>, category <chr>, iow <dbl>, pe_taskid <chr>, maxvmem <dbl>,
#> #   arid <int>, ar_sub_time <dbl>

## Anonymize (although actually already anonymized)
jobs_anon <- anonymize_raw_sge_accounting(jobs)
print(jobs_anon)

#> # A tibble: 1,000 x 45
#>    qname    hostname group  owner   job_name job_number account priority
#>    <chr>    <chr>    <chr>  <chr>   <chr>         <int> <chr>      <int>
#>  1 member.q cin-id3  group1 owner08 sleep.sh          1 sge            0
#>  2 long.q   cc-id2   group1 owner08 sleep.sh          2 sge           19
#>  3 long.q   cc-id2   group1 owner08 sleep.sh          2 sge           19
#>  4 long.q   cc-id2   group1 owner08 sleep.sh          2 sge           19
#>  5 long.q   cc-id2   group1 owner08 sleep.sh          2 sge           19
#>  6 long.q   cc-id2   group1 owner08 sleep.sh          2 sge           19
#>  7 long.q   cc-id2   group1 owner08 sleep.sh          2 sge           19
#>  8 member.q cin-id2  group1 owner08 sleep.sh          2 sge            0
#>  9 long.q   cin-id2  group1 owner08 sleep.sh          2 sge           19
#> 10 long.q   cin-id2  group1 owner08 sleep.sh          2 sge           19
#> # … with 990 more rows, and 37 more variables: submission_time <dbl>,
#> #   start_time <dbl>, end_time <dbl>, failed <int>, exit_status <int>,
#> #   ru_wallclock <dbl>, ru_utime <dbl>, ru_stime <dbl>, ru_maxrss <dbl>,
#> #   ru_ixrss <dbl>, ru_ismrss <dbl>, ru_idrss <dbl>, ru_isrss <dbl>,
#> #   ru_minflt <dbl>, ru_majflt <dbl>, ru_nswap <dbl>, ru_inblock <dbl>,
#> #   ru_oublock <dbl>, ru_msgsnd <dbl>, ru_msgrcv <dbl>, ru_nsignals <dbl>,
#> #   ru_nvcsw <dbl>, ru_nivcsw <dbl>, project <chr>, department <chr>,
#> #   granted_pe <chr>, slots <int>, task_number <int>, cpu <dbl>, mem <dbl>,
#> #   io <dbl>, category <chr>, iow <dbl>, pe_taskid <chr>, maxvmem <dbl>,
#> #   arid <int>, ar_sub_time <dbl>