Index a file for faster access to parts of the file

Index a file for faster access to parts of the file

make_file_index(
  pathname,
  offset = 0,
  skip = 0L,
  n_max = Inf,
  newline = "\n",
  drop_eof = TRUE,
  bfr_size = 5e+07
)

save_file_index(index, file)

read_file_index(file)

Arguments

pathname
(character) The file to be indexed.

offset
(numeric) The number of bytes to skip before start indexing.

skip
(numeric) The number of newline matches to ignore before recording them.

n_max
(numeric) The maximum number of bytes to scan.

newline
(character) The character to scan for.

drop_eof
(logical) If TRUE, the last identified byte offset is dropped if at the very end of the file, i.e. when there is nothing available to read from that position.

bfr_size
(numeric) The number of bytes to read in each iteration.

index
(numeric vector) A sorted index of file byte positions.

file
A pathname to a *.index file to be created or read from.

Value

A numeric vector of file byte offsets that corresponds to the beginning of a line, i.e. a position in the file that was preceeded by a newline character. The first line is at file byte offset 0, which is also always the first element in the returned vector.

Examples

pathname <- system.file("exdata", "accounting", package = "wyntonquery")
index <- read_file_index(sprintf("%s.index", pathname))
cat(sprintf("Number of jobs: %d\n", length(index)))
#> Number of jobs: 1000

## Get all jobs *after* the 500:th job
jobs <- read_sge_accounting(pathname, offset = index[501])
print(jobs)
#> # A tibble: 500 × 45
#>    qname  hostname  group  owner   job_name  job_number account priority
#>    <chr>  <chr>     <chr>  <chr>   <chr>          <int> <chr>      <int>
#>  1 long.q cc-hmid1  group3 owner04 cc-hmid1         238 sge           19
#>  2 long.q cc-id1    group3 owner04 cc-id1           239 sge           19
#>  3 long.q cin-hmid1 group3 owner04 cin-hmid1        242 sge           19
#>  4 long.q qb3-hmid1 group3 owner04 qb3-hmid1        246 sge           19
#>  5 long.q cin-id3   group3 owner04 cin-id3          245 sge           19
#>  6 long.q qb3-id1   group3 owner04 qb3-id1          247 sge           19
#>  7 long.q cin-id2   group3 owner04 cin-id2          244 sge           19
#>  8 long.q qb3-id2   group3 owner04 qb3-id2          248 sge           19
#>  9 long.q cc-id3    group3 owner04 cc-id3           241 sge           19
#> 10 long.q cin-id1   group3 owner04 cin-id1          243 sge           19
#> # … with 490 more rows, and 37 more variables: submission_time <dttm>,
#> #   start_time <dttm>, end_time <dttm>, failed <int>, exit_status <int>,
#> #   ru_wallclock <drtn>, ru_utime <drtn>, ru_stime <drtn>, ru_maxrss <chr>,
#> #   ru_ixrss <chr>, ru_ismrss <chr>, ru_idrss <chr>, ru_isrss <chr>,
#> #   ru_minflt <dbl>, ru_majflt <dbl>, ru_nswap <dbl>, ru_inblock <dbl>,
#> #   ru_oublock <dbl>, ru_msgsnd <dbl>, ru_msgrcv <dbl>, ru_nsignals <dbl>,
#> #   ru_nvcsw <dbl>, ru_nivcsw <dbl>, project <chr>, department <chr>, …

## Get jobs 301 to 350
jobs <- read_sge_accounting(pathname, offset = index[301], n_max = 50)
print(jobs)
#> # A tibble: 50 × 45
#>    qname    hostname group  owner   job_name         job_number account priority
#>    <chr>    <chr>    <chr>  <chr>   <chr>                 <int> <chr>      <int>
#>  1 member.q qb3-id1  group4 owner09 sleep.sh                  2 sge            0
#>  2 long.q   cc-id3   group6 owner01 AD_196_bvFTD_65…          3 sge           19
#>  3 long.q   qb3-id1  group6 owner01 AD_196_bvFTD_65…          4 sge           19
#>  4 long.q   cc-id2   group6 owner01 AD_196_PSP_58_m…          7 sge           19
#>  5 long.q   cin-id3  group6 owner01 bvFTD_65_CONTRO…          9 sge           19
#>  6 long.q   cc-id1   group6 owner01 AD_196_svPPA_46…          8 sge           19
#>  7 long.q   cin-id2  group6 owner01 bvFTD_65_nfvPPA…         10 sge           19
#>  8 long.q   cc-hmid1 group6 owner01 AD_196_CONTROL_…          5 sge           19
#>  9 long.q   cc-id3   group6 owner01 CONTROL_412_PSP…         14 sge           19
#> 10 long.q   cin-id1  group6 owner01 bvFTD_65_PSP_58…         11 sge           19
#> # … with 40 more rows, and 37 more variables: submission_time <dttm>,
#> #   start_time <dttm>, end_time <dttm>, failed <int>, exit_status <int>,
#> #   ru_wallclock <drtn>, ru_utime <drtn>, ru_stime <drtn>, ru_maxrss <chr>,
#> #   ru_ixrss <chr>, ru_ismrss <chr>, ru_idrss <chr>, ru_isrss <chr>,
#> #   ru_minflt <dbl>, ru_majflt <dbl>, ru_nswap <dbl>, ru_inblock <dbl>,
#> #   ru_oublock <dbl>, ru_msgsnd <dbl>, ru_msgrcv <dbl>, ru_nsignals <dbl>,
#> #   ru_nvcsw <dbl>, ru_nivcsw <dbl>, project <chr>, department <chr>, …