Index a file for faster access to parts of the file

Index a file for faster access to parts of the file

make_file_index(
  pathname,
  offset = 0,
  skip = 0L,
  n_max = Inf,
  newline = "\n",
  drop_eof = TRUE,
  bfr_size = 5e+07
)

save_file_index(index, file)

read_file_index(file)

Arguments

pathname

(character) The file to be indexed.

offset

(numeric) The number of bytes to skip before start indexing.

skip

(numeric) The number of newline matches to ignore before recording them.

n_max

(numeric) The maximum number of bytes to scan.

newline

(character) The character to scan for.

drop_eof

(logical) If TRUE, the last identified byte offset is dropped if at the very end of the file, i.e. when there is nothing available to read from that position.

bfr_size

(numeric) The number of bytes to read in each iteration.

index

(numeric vector) A sorted index of file byte positions.

file

A pathname to a *.index file to be created or read from.

Value

A numeric vector of file byte offsets that corresponds to the beginning of a line, i.e. a position in the file that was preceeded by a newline character. The first line is at file byte offset 0, which is also always the first element in the returned vector.

Examples

pathname <- system.file("exdata", "accounting", package = "wyntonquery")
index <- read_file_index(sprintf("%s.index", pathname))
cat(sprintf("Number of jobs: %d\n", length(index)))
#> Number of jobs: 1000

## Get all jobs *after* the 500:th job
jobs <- read_sge_accounting(pathname, offset = index[501])
print(jobs)
#> # A tibble: 500 × 45
#>    qname hostn…¹ group owner job_n…² job_n…³ account prior…⁴ submission_time    
#>    <chr> <chr>   <chr> <chr> <chr>     <int> <chr>     <int> <dttm>             
#>  1 long… cc-hmi… grou… owne… cc-hmi…     238 sge          19 2017-09-05 05:06:49
#>  2 long… cc-id1  grou… owne… cc-id1      239 sge          19 2017-09-05 05:06:49
#>  3 long… cin-hm… grou… owne… cin-hm…     242 sge          19 2017-09-05 05:06:49
#>  4 long… qb3-hm… grou… owne… qb3-hm…     246 sge          19 2017-09-05 05:06:49
#>  5 long… cin-id3 grou… owne… cin-id3     245 sge          19 2017-09-05 05:06:49
#>  6 long… qb3-id1 grou… owne… qb3-id1     247 sge          19 2017-09-05 05:06:49
#>  7 long… cin-id2 grou… owne… cin-id2     244 sge          19 2017-09-05 05:06:49
#>  8 long… qb3-id2 grou… owne… qb3-id2     248 sge          19 2017-09-05 05:06:49
#>  9 long… cc-id3  grou… owne… cc-id3      241 sge          19 2017-09-05 05:06:49
#> 10 long… cin-id1 grou… owne… cin-id1     243 sge          19 2017-09-05 05:06:49
#> # … with 490 more rows, 36 more variables: start_time <dttm>, end_time <dttm>,
#> #   failed <int>, exit_status <int>, ru_wallclock <drtn>, ru_utime <drtn>,
#> #   ru_stime <drtn>, ru_maxrss <chr>, ru_ixrss <chr>, ru_ismrss <chr>,
#> #   ru_idrss <chr>, ru_isrss <chr>, ru_minflt <dbl>, ru_majflt <dbl>,
#> #   ru_nswap <dbl>, ru_inblock <dbl>, ru_oublock <dbl>, ru_msgsnd <dbl>,
#> #   ru_msgrcv <dbl>, ru_nsignals <dbl>, ru_nvcsw <dbl>, ru_nivcsw <dbl>,
#> #   project <chr>, department <chr>, granted_pe <chr>, slots <int>, …

## Get jobs 301 to 350
jobs <- read_sge_accounting(pathname, offset = index[301], n_max = 50)
print(jobs)
#> # A tibble: 50 × 45
#>    qname hostn…¹ group owner job_n…² job_n…³ account prior…⁴ submission_time    
#>    <chr> <chr>   <chr> <chr> <chr>     <int> <chr>     <int> <dttm>             
#>  1 memb… qb3-id1 grou… owne… sleep.…       2 sge           0 2017-08-15 19:00:23
#>  2 long… cc-id3  grou… owne… AD_196…       3 sge          19 2017-08-16 20:51:58
#>  3 long… qb3-id1 grou… owne… AD_196…       4 sge          19 2017-08-16 20:53:24
#>  4 long… cc-id2  grou… owne… AD_196…       7 sge          19 2017-08-16 20:53:24
#>  5 long… cin-id3 grou… owne… bvFTD_…       9 sge          19 2017-08-16 20:53:24
#>  6 long… cc-id1  grou… owne… AD_196…       8 sge          19 2017-08-16 20:53:24
#>  7 long… cin-id2 grou… owne… bvFTD_…      10 sge          19 2017-08-16 20:53:24
#>  8 long… cc-hmi… grou… owne… AD_196…       5 sge          19 2017-08-16 20:53:24
#>  9 long… cc-id3  grou… owne… CONTRO…      14 sge          19 2017-08-16 20:53:24
#> 10 long… cin-id1 grou… owne… bvFTD_…      11 sge          19 2017-08-16 20:53:24
#> # … with 40 more rows, 36 more variables: start_time <dttm>, end_time <dttm>,
#> #   failed <int>, exit_status <int>, ru_wallclock <drtn>, ru_utime <drtn>,
#> #   ru_stime <drtn>, ru_maxrss <chr>, ru_ixrss <chr>, ru_ismrss <chr>,
#> #   ru_idrss <chr>, ru_isrss <chr>, ru_minflt <dbl>, ru_majflt <dbl>,
#> #   ru_nswap <dbl>, ru_inblock <dbl>, ru_oublock <dbl>, ru_msgsnd <dbl>,
#> #   ru_msgrcv <dbl>, ru_nsignals <dbl>, ru_nvcsw <dbl>, ru_nivcsw <dbl>,
#> #   project <chr>, department <chr>, granted_pe <chr>, slots <int>, …